Fix Gem files
